Finished GetInterface(), added _create_apdu(), _convert_to_unicode(), and _send_apdu()
GetInterface() should be up to spec r0.4.

Added _create_apdu() to accept A, INS, P1, P2, L to create a list of the
same data.

Added _send_apdu() primarily for debuging purposes. If debug=True when
MaximAPI is instantiated, the API will print out communication strings
in hex betwen the board and host.

_convert_to_unicode() simply conversts lists of bytes into the
corresponding unicode strings and returns it.